metricas-videos/requirements.txt
ffmpeg-python==0.2.0
xlsxwriter==3.2.9